dedupe windows command execution helpers
